+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next release (tbd)
+
+ - fix: add timeout of 2.5s when polling and log error as JSON
+
 ## v1.1.0 (2021-06-23)
 
  - feat: add configuration to invert status mapping (#1)